requirements.txt
streamlit
streamlit-authenticator
pandas
PyGithub